+++ crates/full-node/sov-stf-runner/src/processes/op_manager/attestations.rs
@@ -59,7 +59,9 @@ where
                 initial_state_root: witness.initial_state_root,
                 slot_hash: witness.da_block_header.hash(),
                 post_state_root: witness.final_state_root,
-                proof_of_bond: self.bonding_proof_service.get_bonding_proof(height),
+                proof_of_bond: self.bonding_proof_service.get_bonding_proof(height).ok_or(
+                    anyhow::anyhow!("Cannot get bonding proof, storage is corrupted."),
+                )?,
             };
 
             let serialized_attestation = SerializedAttestation::from_attestation(&attestation)?;